+++ com/hahn/basic/intermediate/objects/types/TypeDblLike.java
@@ -8,6 +8,6 @@ public class TypeDblLike extends Type {
     
     @Override
     public boolean doesExtend(Type t) {
-        return t == Type.DBL || t == Type.NUMERIC;
+        return t == Type.FLOAT || t == Type.NUMERIC;
     }
 }
